requirements.txt
pydantic>=2.0.0
google-genai>=1.0.0
ddgs>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
google-cloud-texttospeech>=2.0.0

# Additional libraries discovered from imports
numpy>=1.26.0
opencv-python>=4.7.0
Pillow>=10.0.0
moviepy>=1.0.3
edge-tts>=0.14.0
streamlit>=1.34.0
imageio-ffmpeg>=0.4.10